Big commit.  Proof of concept works.
Uses spotify's search api to search for tracks, albums, or artists.
Prompts you for artists (more to come).  Plays selected artist.

I think the architecture is self explanatory, but just in case.
Spotifyr is the main app.  It's job is to ask you what to play and play
it.

One of Spotifyr's args is menu.  Right now I'm using a wrapper I wrote
around dmenu called esmenu.  The important part is that it's constructed
with a list and calling `get()` on it gets an item from the list.  API
subject to change.

Spotify-player-bridge is where I'm controlling the spotify client.
Right now it's using the shell's qdbus and string interpolation.  There
are plenty of dbus clients out there, I just didn't want to get hung up
on that.